% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currently-package.R
\docType{data}
\name{arena}
\alias{arena}
\title{Particle tracking arena}
\description{
Example data from 'BioMassTracking' package.
}
\details{
Comprised of a list with components
\itemize{
\item[lon] longitude vector
\item[lat] latitude vector
\item[U] matrix of U velocity component
\item[V] matrix of V velocity component
\item[S] matrix of S identifiers of regions
}
Orientation of the matrices is transpose to the `image` convention.
}
\examples{
image(arena$lon, arena$lat, t(sqrt(arena$U ^2 + arena$V ^2)))
title("Kergulen Plateau, magnitude of surface currents")
}
