% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currently-package.R
\docType{package}
\name{currently-package}
\alias{currently}
\alias{currently-package}
\title{currently: Particle Tracing Using 'raadtools' Surface Current Vector Fields}
\description{
Particle tracing with current vectors.
}
\author{
\strong{Maintainer}: Michael Sumner \email{mdsumner@gmail.com}

Other contributors:
\itemize{
  \item Thorsten Lenser (wrote original particle code in https://github.com/AustralianAntarcticDivision/BioMassTracking) [contributor]
}

}
\keyword{internal}
