% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/particle.R
\name{particle_trace}
\alias{particle_trace}
\title{Particle tracing}
\usage{
particle_trace(
  xy,
  time_step = 24 * 3600,
  start_date = NULL,
  end_date = NULL,
  method = "bilinear",
  plot = FALSE,
  silent = FALSE
)
}
\arguments{
\item{xy}{input point/s in longitude,latitude}

\item{time_step}{duration of model time step (positive, in seconds)}

\item{start_date}{time to start the model (can be of type Date but beware doing your own math mixing Date/POSIXct)}

\item{end_date}{time to end the model (may be prior to or preceding start_date)}

\item{method}{in the raster lookup, may be 'bilinear' or 'simple'}

\item{plot}{plot the points or not (makes it slower)}

\item{silent}{suppress any message}
}
\description{
Create traces of particles from velocity vector fields. This assumes availability of the raadtools package.
}
\details{
Relies on polar versions of Copernicus altimetry surface velocity fields.

This function is the high-level vectorized "use the raster package" version of some older C-based particle
tracing code. The technique uses raster to extract the x and y displacements in surface current vectors, and
apply those to input points at the given time step. Runge-Kutta is used to generate 4 coefficients for the
displacements.

We can use 'start_date' and 'end_date' to specify the direction, the start is the time we begin the
simulation at, and the end when it finishes. If the 'end_date' is prior to the 'start_date' the time step
is negated internally.
}
