% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rk-package.R
\docType{package}
\name{rk-package}
\alias{rk}
\alias{rk-package}
\title{rk: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
