% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currently-package.R
\name{rk0}
\alias{rk0}
\title{Runge Kutta for particle tracking}
\usage{
rk0(x, ...)
}
\arguments{
\item{x}{x}

\item{...}{dots}
}
\value{
1
}
\description{
Runge Kutta for particle tracking
}
\examples{
rk()
}
