% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rk-package.R
\docType{package}
\name{rk-package}
\alias{rk}
\alias{rk-package}
\title{rk: Particle Tracing Using 'raadtools' Surface Current Vector Fields}
\description{
Particle tracing with current vectors.
}
\author{
\strong{Maintainer}: Michael Sumner \email{mdsumner@gmail.com}

}
\keyword{internal}
